Hide App.DB and create method to reveal
Teh database had to be created in main() to ensure the logger is made
first. This means the tests fail because they don't run App.main(). To
fix this the DB is not private and there is a method to create the
database if it is null and then reveal the database